refactor: extract shared effect handlers into std/caps crate
