+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "examples/discovery",
 ]
 exclude = [
+    "std/caps",          # WASI guest lib — targets wasm32-wasip2, not host
     "std/shell",         # WASI guest — targets wasm32-wasip2, not host
     "examples/echo",     # WASI guest — targets wasm32-wasip2, not host
     "examples/counter",  # WASI guest — targets wasm32-wasip2, not host
