+++ std/caps/build.rs
@@ -1,0 +1,35 @@
+use std::env;
+use std::path::Path;
+
+/// Build script for the caps crate.
+///
+/// Compiles shared system schemas needed by the effect handler factories.
+fn main() {
+    let manifest_dir = env::var("CARGO_MANIFEST_DIR").expect("CARGO_MANIFEST_DIR not set");
+    let manifest_path = Path::new(&manifest_dir);
+
+    let capnp_dir = manifest_path
+        .join("../..")
+        .join("capnp")
+        .canonicalize()
+        .expect("capnp dir not found");
+
+    // Shared schemas needed by effect handlers
+    capnpc::CompilerCommand::new()
+        .src_prefix(&capnp_dir)
+        .file(capnp_dir.join("system.capnp"))
+        .file(capnp_dir.join("ipfs.capnp"))
+        .file(capnp_dir.join("routing.capnp"))
+        .file(capnp_dir.join("stem.capnp"))
+        .file(capnp_dir.join("http.capnp"))
+        .run()
+        .expect("failed to compile shared capnp schemas");
+
+    // Cargo rebuild triggers
+    for schema in &["system", "ipfs", "routing", "stem", "http"] {
+        println!(
+            "cargo:rerun-if-changed={}",
+            capnp_dir.join(format!("{schema}.capnp")).display()
+        );
+    }
+}
